@@ -43,7 +43,7 @@
     "commitizen": "^3.1.1",
     "cz-conventional-changelog": "^2.1.0",
     "enquirer": "^2.3.0",
-    "eslint": "^4.19.1",
+    "eslint": "^7.5.0",
     "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-prettier": "^3.1.2",
     "esm": "^3.2.25",
